Update memory load function to use uint16_t's

+++ include/memory.hpp
@@ -4,7 +4,7 @@
 class Memory {
     public:
         Memory();
-        void load(std::vector<uint8_t> data, int address);
+        void load(std::vector<uint8_t> data, uint16_t address);
     private:
         std::vector<uint8_t> memory;
     friend class CPU;
